requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
djangorestframework==3.15.1
et-xmlfile==1.1.0
install==1.3.5
numpy==1.24.4
openpyxl==3.1.2
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
django-environ==0.9.0
environ==1.0
psycopg2-binary==2.9.9
